app/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-multipart==0.0.6
python-dotenv==0.21.1
beautifulsoup4==4.11.2
requests==2.31.0  # Instead of aiohttp
redis==4.5.1
httpx==0.24.1